requirements.txt
allure-pytest==2.14.3
allure-python-commons==2.14.3
attrs==25.3.0
certifi==2025.6.15
cfgv==3.4.0
charset-normalizer==3.4.2
distlib==0.3.9
execnet==2.1.1
filelock==3.18.0
flake8==7.3.0
greenlet==3.2.3
identify==2.6.12
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mccabe==0.7.0
nodeenv==1.9.1
packaging==25.0
platformdirs==4.3.8
playwright==1.52.0
pluggy==1.6.0
pre_commit==4.2.0
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pyee==13.0.0
pyflakes==3.4.0
Pygments==2.19.2
pytest==8.4.1
pytest-base-url==2.1.0
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-playwright==0.7.0
pytest-rerunfailures==15.1
pytest-xdist==3.7.0
python-dotenv==1.1.0
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.4
text-unidecode==1.3
typing_extensions==4.14.0
urllib3==2.5.0
virtualenv==20.31.2
